requirements.txt
Flask
tensorflow
numpy
newspaper3k
scikit-learn
requests
beautifulsoup4
lxml
gunicorn
